varios cambios

+++ src/sys/DB-Engine/links/followTweet.ts
@@ -1,7 +1,6 @@
 import { error } from "../../logging";
 import getPool from "../database";
 
-// Mejorado: PascalCase para la interfaz
 interface FollowTweetX {
     id: number;
     guild_id: string;
@@ -11,6 +10,7 @@ interface FollowTweetX {
     lang: string | null;
     Domain: string | null;
     addBy: string;
+    onlyMedia: boolean;
     created_at: Date;
 }
 
@@ -18,7 +18,7 @@ interface FollowTweetX {
 export async function getAllFollowTweet(): Promise<FollowTweetX[] | null> {
     try {
         const pool = await getPool();
-        const [rows] = await pool.query("SELECT id, guild_id, canal, xUser, lastPost, lang, customDomain, addBy, created_at FROM followTweetX");
+        const [rows] = await pool.query("SELECT id, guild_id, canal, xUser, lastPost, lang, customDomain, addBy, onlyMedia, created_at FROM followTweetX");
         return rows as FollowTweetX[];
     } catch (e) {
         error(`Falló la recuperación de datos de followTweet: ${e}`);
@@ -29,7 +29,8 @@ export async function getAllFollowTweet(): Promise<FollowTweetX[] | null> {
 export async function updateFollowTweet(gremio: string, canal: string, xUser: string, lastPost: string): Promise<boolean> {
     try {
         const pool = await getPool();
-        await pool.query(`UPDATE followTweetX SET lastPost = ? WHERE guild_id = ? AND canal = ? AND xUser = ?`, [lastPost, gremio, canal, xUser]);
+        await pool.query(`UPDATE followTweetX SET lastPost = ? WHERE guild_id = ? AND canal = ? AND xUser = ?`,
+            [lastPost, gremio, canal, xUser]);
         return true;
     } catch (e) {
         error(`Falló la actualización de datos de followTweet: ${e}`);
@@ -49,23 +50,26 @@ export async function deleteFollowTweetByGuild(guild: string): Promise<boolean> 
 }
 
 /// ============================ BD4Comands ============================ ///
-
-export async function addFollowTweet(dta: FollowTweetX): Promise<{ success: boolean; message: string }> {
+export async function addFollowTweet(dta: FollowTweetX): Promise<boolean> {
     try {
         const pool = await getPool();
-        await pool.query("INSERT INTO followTweetX (guild_id, canal, xUser, lang, customDomain, addBy) VALUES (?, ?, ?, ?, ?, ?)", [dta.guild_id, dta.canal, dta.xUser, dta.lang, dta.Domain, dta.addBy]);
-        return { success: true, message: "Se añadio correctamente!!" };
-    } catch (e: any) {
-        if (e.code === 'ER_DUP_ENTRY') { return { success: false, message: "Este usuario ya está siendo seguido en el servidor." }; }
-        error(`Falló la adición de datos de followTweet: ${e}`);
-        return { success: false, message: "Algo salio mal!!" };
+        await pool.query(`
+            INSERT INTO followTweetX (guild_id, canal, xUser, lang, customDomain, addBy, onlyMedia) VALUES (?, ?, ?, ?, ?, ?, ?) 
+            ON DUPLICATE KEY UPDATE canal = ?, lang = ?, customDomain = ?, onlyMedia = ?`,
+            [dta.guild_id, dta.canal, dta.xUser, dta.lang, dta.Domain, dta.addBy, dta.onlyMedia,
+            dta.canal, dta.lang, dta.Domain, dta.onlyMedia]);
+        return true;
+    } catch (e) {
+        error(`Falló la operación en followTweet: ${e}`);
+        return false;
     }
 }
 
 export async function followTweetonGuild(gremio: string): Promise<FollowTweetX[] | null> {
     try {
         const pool = await getPool();
-        const [rows] = await pool.query("SELECT id, guild_id, canal, xUser, lastPost, lang, customDomain, addBy, created_at FROM followTweetX WHERE guild_id = ?", [gremio]);
+        const [rows] = await pool.query("SELECT id, guild_id, canal, xUser, lastPost, lang, customDomain, addBy, onlyMedia, created_at FROM followTweetX WHERE guild_id = ?",
+            [gremio]);
         return rows as FollowTweetX[];
     } catch (e) {
         error(`Falló la recuperación de datos de followTweet en guild: ${e}`);
@@ -73,10 +77,14 @@ export async function followTweetonGuild(gremio: string): Promise<FollowTweetX[]
     }
 }
 
-export async function deleteFollowTweet(gremio: string, id: number): Promise<boolean> {
+interface deleteFollowInt { gremio: string, id?: number, xUser?: string }
+export async function deleteFollowTweet(dta: deleteFollowInt): Promise<boolean> {
+    const { gremio, id, xUser } = dta;
     try {
+        if (!id && !xUser) return false;
         const pool = await getPool();
-        await pool.query(`DELETE FROM followTweetX WHERE guild_id = ? AND id = ?`, [gremio, id]);
+        if (id) await pool.query(`DELETE FROM followTweetX WHERE guild_id = ? AND id = ?`, [gremio, id]);
+        else await pool.query(`DELETE FROM followTweetX WHERE guild_id = ? AND xUser = ?`, [gremio, xUser]);
         return true;
     } catch (e) {
         error(`Falló la eliminación de datos de followTweet (ID): ${e}`);
